+++ careers/server/gameManager.py
@@ -31,28 +31,6 @@ class Game(BaseModel):
     joinCode: str = Field(...)
     inProgress: bool = Field(default=False)
 
-class User(BaseModel):
-    name: str = Field(...)
-    email: str = Field(...)
-    initials: str = Field(...)
-    id: str = Field(alias="_id", default=None)
-    createdDate: datetime = Field(default=datetime.now())
-
-class CareersUserManager(object):
-
-    def __init__(self):
-        self.config = dotenv.dotenv_values(".env")
-        self.mongo_client = MongoClient(self.config["DB_URL"])
-        self.database = self.mongo_client["careers"]
-        self.database["users"].create_index('name')
-        self.collection = self.database["users"]
-
-    def createUser(self, user: User) -> User:
-        user.id = str(uuid4())
-        self.collection.insert_one(jsonable_encoder(user))
-
-        return user
-
 class CareersGameManager(object):
 
     def __init__(self):
